refactor: listing/trade 도메인 패키지 구조를 다른 도메인과 통일
