+++ Pokade/src/main/java/com/pokade/domain/listing/entity/ListingGrade.java
@@ -1,4 +1,4 @@
-package com.pokade.domain.listing;
+package com.pokade.domain.listing.entity;
 
 public enum ListingGrade {
     S,
