+++ Pokade/src/main/java/com/pokade/domain/listing/repository/ListingRepository.java
@@ -1,5 +1,7 @@
-package com.pokade.domain.listing;
+package com.pokade.domain.listing.repository;
 
+import com.pokade.domain.listing.entity.Listing;
+import com.pokade.domain.listing.entity.ListingStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -27,7 +29,7 @@ public interface ListingRepository extends JpaRepository<Listing, Long> {
 
     // ACTIVE 상태인 매물만 원자적으로 TRADING으로 전환. 반환값 0 = 이미 팔렸거나 존재하지 않음(동시 구매 충돌)
     @Modifying
-    @Query("UPDATE Listing l SET l.status = com.pokade.domain.listing.ListingStatus.TRADING "
-            + "WHERE l.id = :id AND l.status = com.pokade.domain.listing.ListingStatus.ACTIVE")
+    @Query("UPDATE Listing l SET l.status = com.pokade.domain.listing.entity.ListingStatus.TRADING "
+            + "WHERE l.id = :id AND l.status = com.pokade.domain.listing.entity.ListingStatus.ACTIVE")
     int markAsTrading(@Param("id") Long listingId);
 }
