+++ Pokade/src/main/java/com/pokade/domain/trade/entity/PaymentMethod.java
@@ -1,4 +1,4 @@
-package com.pokade.domain.trade;
+package com.pokade.domain.trade.entity;
 
 public enum PaymentMethod {
     CARD,
