+++ Pokade/src/main/java/com/pokade/domain/trade/service/TradeService.java
@@ -1,9 +1,14 @@
-package com.pokade.domain.trade;
+package com.pokade.domain.trade.service;
 
-import com.pokade.domain.listing.Listing;
-import com.pokade.domain.listing.ListingRepository;
+import com.pokade.domain.listing.entity.Listing;
+import com.pokade.domain.listing.repository.ListingRepository;
 import com.pokade.domain.trade.dto.TradeCreateRequest;
 import com.pokade.domain.trade.dto.TradeResponse;
+import com.pokade.domain.trade.entity.Payment;
+import com.pokade.domain.trade.entity.PaymentMethod;
+import com.pokade.domain.trade.entity.Trade;
+import com.pokade.domain.trade.repository.PaymentRepository;
+import com.pokade.domain.trade.repository.TradeRepository;
 import com.pokade.global.exception.BusinessException;
 import com.pokade.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
